// Front/src/types/Github/Milestone.ts
import { z } from "zod";

export const MilestoneSchema = z.object({
    url: z.string(),
    html_url: z.string(),
    labels_url: z.string(),
    id: z.number(),
    node_id: z.string(),
    number: z.number(),
    title: z.string(),
    description: z.string(),
    creator: z.object({
        login: z.string(),
        id: z.number(),
        node_id: z.string(),
        avatar_url: z.string(),
        gravatar_id: z.string(),
        url: z.string(),
        html_url: z.string(),
        followers_url: z.string(),
        following_url: z.string(),
        gists_url: z.string(),
        starred_url: z.string(),
        subscriptions_url: z.string(),
        organizations_url: z.string(),
        repos_url: z.string(),
        events_url: z.string(),
        received_events_url: z.string(),
        type: z.string(),
        user_view_type: z.string(),
        site_admin: z.boolean(),
    }),
    open_issues: z.number(),
    closed_issues: z.number(),
    state: z.string(),
    created_at: z.string(),
    updated_at: z.string(),
    due_on: z.coerce.date(),
    closed_at: z.string().nullable(),
});

export type MilestoneType = z.infer<typeof MilestoneSchema>;